pyproject.toml
[project]
name = "ieor4576-project2"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai-agents>=0.0.15",
    "litellm>=1.56.0",
    "google-cloud-aiplatform>=1.77.0",
    "duckdb>=1.2.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "matplotlib>=3.9.0",
    "plotly>=5.24.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "yfinance>=0.2.40",
    "aiofiles>=24.1.0",
    "kaleido>=0.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "tools", "models"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
